Revert "add first purchase discount to courseware (#22058)"
This reverts commit 059b6197945d01758cfb0e831333a588425b8876.
