+++ lms/static/sass/_build-lms-v1.scss
@@ -74,7 +74,6 @@
 @import 'features/content-type-gating';
 @import 'features/course-duration-limits';
 @import 'features/enterprise-learner-portal-banner';
-@import 'features/first-purchase-banner';
 
 // search
 @import 'search/search';
