+++ lms/static/sass/bootstrap/lms-main.scss
@@ -26,7 +26,7 @@ $static-path: '../..';
 @import 'features/course-upgrade-message';
 @import 'features/course-duration-limits';
 @import 'features/enterprise-learner-portal-banner';
-@import 'features/first-purchase-banner';
+
 
 // Individual Pages
 @import "views/program-marketing-page";
