+++ lms/static/sass/features/_course-experience.scss
@@ -20,6 +20,26 @@
   }
 }
 
+// First purchase offer banner
+.first-purchase-offer-banner {
+  background-color: #DEE3F1;
+  font-size: 16px;
+  border-radius: 7px;
+  padding: 20px;
+  margin: 20px auto;
+
+  .first-purchase-offer-banner-bold {
+      font-weight: bold;
+      color: #23419F;
+  }
+
+  a {
+    color: #00496f;
+    text-decoration: underline;
+    font-weight: bold;
+  }
+}
+
 // Course call to action message
 .course-message {
   display: flex;
