--- lms/static/sass/features/_first-purchase-banner.scss
@@ -1,27 +1,0 @@
-// First purchase offer banner
-.first-purchase-offer-banner {
-  background-color: #DEE3F1;
-  font-size: 16px;
-  border-radius: 7px;
-  padding: 20px;
-  margin: 20px auto;
-  box-sizing: border-box;
-  line-height: 1.5;
-
-  .first-purchase-offer-banner-bold {
-      font-weight: bold;
-      color: #393f43;
-  }
-
-  a {
-    color: #23419F !important;
-    text-decoration: underline !important;
-    font-weight: bold !important;
-    border-bottom: none;
-  }
-
-}
-
-#seq_content .first-purchase-offer-banner {
-  max-width: $text-width-readability-max;
-}
