+++ lms/static/sass/views/_verification.scss
@@ -2265,11 +2265,11 @@
 
     .first-purchase-offer-banner-bold {
       font-weight: bold;
-      color: #393f43;
+      color: #23419f;
     }
 
     a {
-      color: #23419F;
+      color: #00496f;
       text-decoration: underline !important;
       font-weight: bold;
       border-bottom: none;
